+++ covalent_dispatcher/entry_point.py
@@ -27,6 +27,8 @@ from concurrent.futures import ThreadPoolExecutor
 
 from covalent._results_manager.result import Result
 from covalent._shared_files import logger
+from covalent._shared_files.config import get_config
+from covalent._workflow.transport import _TransportGraph
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info